api/requirements.txt
Flask==2.3.2
neo4j==5.12.0
psycopg2-binary==2.9.6
pandas==2.0.3
numpy==1.23.5
